+++ sever/src/main/kotlin/shareApp/ShareAppApplication.kt
@@ -1,37 +1,34 @@
 package shareApp
 
-import spark.Spark.*
-import shareApp.controller.*
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+import shareApp.conf.Routes
 import shareApp.db.dbConnect
-
 
 fun main(args: Array<String>) {
 
-    val toJson = JsonTransformer(ObjectMapper().registerKotlinModule())
+
     dbConnect() //データベース接続
     filter() //cors許容
+    Routes().init()
 
-    path("/users") {
-        get("", UserController().getUser(), toJson)
-        get("/me", UserController().getUserMe(), toJson)
-        get("/groups/:id", UserController().getUserList(), toJson)
-        get("/groups",UserController().getGroups(), toJson)
-        post("", UserController().addUser(), toJson)
-    }
-
-    path("/login") {
-        post("", UserController().login(), toJson)
-    }
-
-    path("groups") {
-        post("", GroupController().addGroup(), toJson)
-    }
-
-    path("/tasks") {
-        get("/:id", TaskController().getTask(), toJson)
-        get("", TaskController().getTaskList(), toJson)
-        post("", TaskController().addTask(), toJson)
-    }
+//    path("/users") {
+//        get("", UserController().getUser(), toJson)
+//        get("/me", UserController().getUserMe(), toJson)
+//        get("/groups/:id", UserController().getUserList(), toJson)
+//        get("/groups",UserController().getGroups(), toJson)
+//        post("", UserController().addUser(), toJson)
+//    }
+//
+//    path("/login") {
+//        post("", UserController().login(), toJson)
+//    }
+//
+//    path("groups") {
+//        post("", GroupController().addGroup(), toJson)
+//    }
+//
+//    path("/tasks") {
+//        get("/:id", TaskController().getTask(), toJson)
+//        get("", TaskController().getTaskList(), toJson)
+//        post("", TaskController().addTask(), toJson)
+//    }
 }
